+++ cubical/DefaultParset.cfg
@@ -39,7 +39,7 @@ ddid-density      = 0.5
 [sol]
 _Help			 = Solution options
 precision        = 32            # Solve in single or double precision #options:32|64
-jones-type       = complex-2x2   # Type of Jones matrix to solve for #options:complex-2x2|phase-diag
+jones-type       = complex-2x2   # Type of Jones matrix to solve for #options:complex-2x2|phase-diag|robust-2x2
 time-int         = 1
 freq-int         = 1
 max-iter         = 200
