+++ cubical/solver.py
@@ -3,12 +3,13 @@ Implements the solver loop
 """
 import numpy as np
 import traceback
-from cubecal.tools import logger, ModColor
-from cubecal.data_handler import FL, Tile
-from cubecal.tools import shared_dict
-from cubecal.machines import complex_2x2_machine
-from cubecal.machines import phase_diag_machine
-from cubecal.statistics import SolverStats
+from cubical.tools import logger, ModColor
+from cubical.data_handler import FL, Tile
+from cubical.tools import shared_dict
+from cubical.machines import complex_2x2_machine
+from cubical.machines import complex_W_2x2_machine
+from cubical.machines import phase_diag_machine
+from cubical.statistics import SolverStats
 
 log = logger.getLogger("solver")
 
@@ -78,6 +79,8 @@ def _solve_gains(obser_arr, model_arr, flags_arr, options, label="", compute_res
         gm = complex_2x2_machine.Complex2x2Gains(model_arr, options)
     elif options['jones-type'] == 'phase-diag':
         gm = phase_diag_machine.PhaseDiagGains(model_arr, options)
+    if options['jones-type'] == 'robust-2x2':
+        gm = complex_W_2x2_machine.ComplexW2x2Gains(model_arr, options)
     else:
         raise ValueError("unknown jones-type '{}'".format(options['jones-type']))
