+++ cubical/tools/NpShared.py
