+++ setup.py
@@ -34,25 +34,25 @@ class custom_install(install):
 		install.run(self)
 
 extensions = [Extension(
-						"cubecal.kernels.cyfull_complex", ["cubecal/kernels/cyfull_complex.pyx"],
+						"cubical.kernels.cyfull_complex", ["cubical/kernels/cyfull_complex.pyx"],
         				include_dirs=[np.get_include()], extra_compile_args=['-fopenmp', 
         				'-ffast-math', '-O2', '-march=native',	'-mtune=native', '-ftree-vectorize'],
 						extra_link_args=['-lgomp']
 					   ),
 			        Extension(
-						"cubecal.kernels.cyphase_only", ["cubecal/kernels/cyphase_only.pyx"],
+						"cubical.kernels.cyphase_only", ["cubical/kernels/cyphase_only.pyx"],
         				include_dirs=[np.get_include()], extra_compile_args=['-fopenmp', 
         				'-ffast-math', '-O2', '-march=native',	'-mtune=native', '-ftree-vectorize'],
 						extra_link_args=['-lgomp']
 					   ),
               Extension(
-            "cubecal.kernels.cyfull_W_complex", ["cubecal/kernels/cyfull_W_complex.pyx"],
+            "cubical.kernels.cyfull_W_complex", ["cubical/kernels/cyfull_W_complex.pyx"],
                 include_dirs=[np.get_include()], extra_compile_args=['-fopenmp', 
                 '-ffast-math', '-O2', '-march=native',  '-mtune=native', '-ftree-vectorize'],
             extra_link_args=['-lgomp']
              )]
 
-setup(name='cubecal',
+setup(name='cubical',
       version='0.2.1',
       description='Fast calibration implementation exploiting complex optimisation.',
       url='https://github.com/JSKenyon/phd-code',
@@ -69,7 +69,7 @@ setup(name='cubecal',
       author_email='jonosken@gmail.com',
       license='GNU GPL v3',
       cmdclass={'install': custom_install},	 
-      packages=['cubecal', 'cubecal/tools', 'cubecal/kernels', 'cubecal/machines'],
+      packages=['cubical', 'cubical/tools', 'cubical/kernels', 'cubical/machines'],
       setup_requires=['numpy', 'cython'],
       install_requires=[  'numpy', 
                           'cython', 
@@ -78,7 +78,7 @@ setup(name='cubecal',
                           'sharedarray', 
                           'matplotlib'  ],
       include_package_data=True,
-      package_data={'cubecal'	:	[	'main.py',
+      package_data={'cubical'	:	[	'main.py',
                 									'solver.py', 
           		      							'statistics.py',
           		      							'plots.py',
@@ -86,21 +86,21 @@ setup(name='cubecal',
           		      							'MBTiggerSim.py', 
           		      							'data_handler.py',
           		      							'TiggerSourceProvider.py', 
-          		      							'bin/gocubecal'], 
-                    'tools'		:	[	'cubecal/tools/logger.py', 
-		                       			  'cubecal/tools/ModColor.py',
-		                    		 	    'cubecal/tools/ClassPrint.py',
-		                    		 	    'cubecal/tools/myoptparse.py',
-		                    		 	    'cubecal/tools/parsets.py'],
-                    'kernels'	:	[	'cubecal/kernels/cyfull_complex.pyx', 
-                       					  'cubecal/kernels/cyphase_only.pyx',
-                                  'cubecal/kernels/cyfull_W_complex.pyx'],
-                    'machines'	:	[	'cubecal/machines/abstract_gain_machine.py', 
-                       					    'cubecal/machines/complex_2x2_machine.py',
-                       					    'cubecal/machines/phase_diag_machine.py',
-                                    'cubecal/machines/complex_W_2x2_machine.py'] },
+          		      							'bin/gocubical'],
+                    'tools'		:	[	'cubical/tools/logger.py',
+		                       			  'cubical/tools/ModColor.py',
+		                    		 	    'cubical/tools/ClassPrint.py',
+		                    		 	    'cubical/tools/myoptparse.py',
+		                    		 	    'cubical/tools/parsets.py'],
+                    'kernels'	:	[	'cubical/kernels/cyfull_complex.pyx',
+                       					  'cubical/kernels/cyphase_only.pyx',
+                                  'cubical/kernels/cyfull_W_complex.pyx'],
+                    'machines'	:	[	'cubical/machines/abstract_gain_machine.py',
+                       					    'cubical/machines/complex_2x2_machine.py',
+                       					    'cubical/machines/phase_diag_machine.py',
+                                    'cubical/machines/complex_W_2x2_machine.py'] },
       zip_safe=False,
       ext_modules = cythonize(extensions),
-      scripts=['cubecal/bin/gocubecal'],
+      scripts=['cubical/bin/gocubical'],
 )
